requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.4
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
